Includes the whole file object in the include
This allows us to echo key.contents as well as other yaml frontmatter in the included files.
The advantage to this is that it allows us to keep details pertaining to each partial contained in that partials id file.
(Useful for metadata, block level links and the like, tooltip/popover content, etc.)